Distinguish between a `Wallet`'s `Locked` and `Unlocked` states (#540)
* Distinguish between a `Wallet`'s `Locked` and `Unlocked` states

This PR was motivated by work @kayagokalp and I are doing on the
following:
https://github.com/FuelLabs/sway/issues/2413

We noticed that in order to perform any operations using a `Wallet`, the
private key must be known, even if those operations were simply
retrieving asset balances or a list of transactions.

This PR aims to draw a clearer distinction between Wallet methods that
require access to a private key and those that do not. The idea is to
encourage users of the API to be conscious of when their wallet is in an
unlocked state (i.e. their private key resides in memory) and when it
does not.

To achieve this, we turn `Wallet` into a session type, parameterised
over its state which can be either `Locked` or `Unlocked`.

The default state is `Locked` in the case that the parameter is not
specified for `Wallet` or `LocalWallet`, meaning that if users require
holding the wallet in an unlocked state they must specify the type
explicitly, e.g. `Wallet<Unlocked>`. I think this explicitness can
assist us in maintaining awareness of the state of the wallet that we
require in our downstream APIs.

Users should generally prefer using the `Wallet` in a locked state (the
default) and should ideally only `unlock` the wallet for the duration
that they need to perform operations requiring the private key. Users
can now construct a `Wallet` in a `Locked` state, e.g.
`Wallet::from_address(public_addr, provider)` and access the various
useful read-only operations without knowing the private key or having it
stored in memory.

As a follow-up to this work, we should take a closer look at the usage
of `LocalWallet` in downstream crates (including `fuels-contract`) to
see if we can refactor these to use the wallet in a `Locked` state more
often, only moving into the `Unlocked` state when absolutely necessary.

* Update overlooked tests for recent Wallet state API changes

* Update the book with a chapter on Wallet Access (Unlocked vs Locked)

* Fix some repetitive wording in wallet-access chapter

* Move `generate_mnemonic_phrase` from associated to a free function

Previously, the `generate_mnemonic_phrase` function was associated with
the `Wallet` type.

As of introducing the wallet state parameter, the associated function
can no longer be called without filling the state type parameter. E.g.
what could previously be written as `Wallet::generate_mnemonic_phrase`
would now need to be written as
`Wallet::<Locked>::generate_mnemonic_phrase`, even though wallet state
has nothing to do with the function itself.

To avoid these type inference issues, the function has been moved to the
wallet module level. It has been included in the `fuels` prelude for
ease of access.

* Simplify wallet access by removing the state type input parameter

This removes the input state type from `Wallet<T>` in favour of using
two simple types, `Wallet` and `WalletUnlocked`.

`WalletUnlocked` provides a `Deref` implementation that targets its
inner `Wallet`, allowing it to be used with API's that take a `&Wallet`
as an input and to call the `Wallet` methods directly.

The `LocalWallet` has been removed as a part of this commit in order to
unnecessarily introduce a new `LocalWalletUnlocked` type when we may
wish to remove the ambiguous `LocalWallet` type in either way: see #543.

* Update wallet access docs for removal of wallet state types

* Add missing LocalWallet -> WalletUnlocked in fuels-signers

* Update new flat_abigen module for LocalWallet -> WalletUnlocked

* Fix two typos in Wallet doc comment