feat: return Enums containing heap types at depth 1 (#1106)
This PR enables returning heap types surrounded in an `Enum` type. It
notably enables returning a `Result<Vec>` from a contract call.

Closes #940.

Main disavantages from this approach: 
- We are injecting more custom bytecode
- Enum can contain only one variant that uses a heap type inside. It
could be extended to having an arbitrary number of them, but they would
all have to have the same encoding width. This is a personal design
choice. 
Co-authored-by: segfault-magnet <37515857+segfault-magnet@users.noreply.github.com>