feat: experimental encoding for contracts, scripts and predicates (#1303)
closes: https://github.com/FuelLabs/fuels-rs/issues/1278,
https://github.com/FuelLabs/fuels-rs/issues/1279,
https://github.com/FuelLabs/fuels-rs/issues/1046

This PR adds support for the new encoding scheme for contracts, scripts
and predicates.

I have added a new `ExperimentalBoundedEncoder` which can be activated
with the `experimental` cfg flag.
I have tried to minimize the impact of the new encoder as much as
possible to make it easier for review. A full refactor of the whole sdk
is necessary once the new encoding becomes the default one.

- The function selector changed and now it is the name of the method.
- The `CALL` opcode changed with the new encoding and is expecting the
following call data: ContractID, pointer to fn_selector (name of the
method), pointer to encoded arguments, number of coins, asset_id,
gas_forwarded.