feat: switch blob ID calculation from data offset to configurable offset (#1593)
<!--
List the issues this PR closes (if any) in a bullet list format, e.g.:
- Closes #ABCD
- Closes #EFGH
-->

# Release notes

<!--
Use this only if this PR requires a mention in the Release
Notes Summary. Valuable features and critical fixes are good
examples. For everything else, please delete the whole section.
-->

In this release, we:

~~- Previously deployed blobs are no longer accessible due to changes in
blob ID calculation. Redeployment is required.~~


# Summary
Updates blob ID calculation to use the configurable offset instead of
the data offset, aligning with the updated design for scripts and
predicates.

`has_configurable_section_offset` function checks bytes 4 and 7 to
distinguish binary versions. If byte 4 is 0x74 (JMPF), it considers byte
7: 0x02 means legacy, 0x04 means new, and any other value results in an
error.

<!--
Please write a summary of your changes and why you made them.
Not all PRs will be complex or substantial enough to require this
section, so you can remove it if you think it's unnecessary.
-->


# Checklist

- [x] All **changes** are **covered** by **tests** (or not applicable)
- [x] All **changes** are **documented** (or not applicable)
- [x] I **reviewed** the **entire PR** myself (preferably, on GH UI)
- [x] I **described** all **Breaking Changes** (or there's none)

---------

Co-authored-by: hal3e <git@hal3e.io>
Co-authored-by: Ahmed Sagdati <37515857+segfault-magnet@users.noreply.github.com>