chore: update public key recovery example in docs (#1638)
When viewing the [docs hub
example](https://docs.fuel.network/docs/fuels-rs/wallets/signing/#signing)
on how to recover a public key, the variable name provided is
`recovered_address`. This is misleading and leads to confusion as the
recovered type is a `PublicKey`
https://github.com/FuelLabs/fuel-vm/blob/09c500838b734603970c62d414e0e95825f44773/fuel-crypto/src/secp256/signature.rs#L140C56-L140C65.
This has been updated to a more appropriate name.

# Checklist

- [x] All **changes** are **covered** by **tests** (or not applicable)
- [x] All **changes** are **documented** (or not applicable)
- [x] I **reviewed** the **entire PR** myself (preferably, on GH UI)
- [x] I **described** all **Breaking Changes** (or there's none)